Change pathgen to use IR for code generation.
This PR is actually dependent on these PRs in ygot:
* https://github.com/openconfig/ygot/pull/681
* https://github.com/openconfig/ygot/pull/683

In the future, any additional info required from the goyang AST (i.e.  `yang.Entry`) need to first be put in the IR before being able to be used by the code generator. The intention is to separate the concerns of creating a code-generation-friendly hierarchical structure that has information already extracted from the AST, from the actual code generation logic.