Add flag for generating ygot setters.
I made this a flag since there could be conflicts. Also made this
true by default. This might break some users but if we don't make it
default now it will never be.

Alternatively we can make it always `true` but since this is a new flag
in ygot some folks might already be depending on it, so I didn't want to
break it for them.