Panic on GNMI delete with union leafs (#167)
This patch modifies `pathgen/gnmigen.go` to check if the reflect value is valid before calling IsZero(). In the case of YANG union values, an uninitialized interface can be passed to reflect.ValueOf() which is one of the few cases it can return a zero value. This condition seems to exist only when a GNMI message is received that contains enough for a struct to be allocated but not enough for values to be populated (DELETE on a YANG keyed list).

A new test case is added to `ygnmi/ygnmi_test.go` which utilizes the modified YANG schema in `pathgen/testdata/yang/openconfig-unione.yang` to trigger the failure.